requirements.txt
pymilvus==2.6.12
pymilvus.model==0.3.2
langchain-huggingface==1.2.2
langchain-core==1.3.2
langchain==1.2.15
sentence-transformers==5.4.1
transformers==5.6.2
underthesea==9.4.0
underthesea_core==3.3.0
redis==7.4.0
pymongo==4.17.0
fastapi==0.117.1
uvicorn==0.42.0
sse-starlette==3.0.3
python-dotenv==1.1.0
llama_cpp_python==0.3.21
langfuse==4.5.1
numpy==2.3.5
pandas==2.3.3
datasets==4.8.5
ragas==0.4.3
langchain-google-genai==4.2.2
google-genai==1.73.1
langchain-community==0.4.1
